dev.requirements.txt
-r ./requirements.txt
pip
tox
tox-pip-extensions
isort
flake8
factory-boy
pytest
pytest-django
pytest-cov
pytest-mock
pytest-xdist
freezegun
black
django-coverage-plugin

